Elevate AbstractOptimizer and ConstantFoldingOptimizer
