+++ server/storage-jdbc-postgresql/src/main/java/org/bithon/server/storage/jdbc/postgresql/PostgresqlDialect.java
@@ -27,7 +27,7 @@ import org.bithon.component.commons.expression.MapAccessExpression;
 import org.bithon.component.commons.expression.function.Functions;
 import org.bithon.component.commons.expression.function.builtin.AggregateFunction;
 import org.bithon.component.commons.expression.function.builtin.StringFunction;
-import org.bithon.component.commons.expression.optimzer.ExpressionOptimizer;
+import org.bithon.component.commons.expression.optimzer.AbstractOptimizer;
 import org.bithon.component.commons.time.DateTime;
 import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.commons.time.TimeSpan;
@@ -100,7 +100,7 @@ public class PostgresqlDialect implements ISqlDialect {
 
     @Override
     public IExpression transform(IExpression expression) {
-        return expression == null ? null : expression.accept(new ExpressionOptimizer.AbstractOptimizer() {
+        return expression == null ? null : expression.accept(new AbstractOptimizer() {
             /**
              * H2 does not support Map, the JSON formatted string is stored in the column.
              * So we turn the MapAccessExpression into a LIKE expression
